backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
ddgs==9.1.1
distro==1.9.0
duckduckgo_search==8.1.1
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.0
MarkupSafe==3.0.2
openai==1.95.0
primp==0.15.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
